requirements.txt
Flask==2.3.3
pdfplumber==0.10.3
PyMuPDF==1.23.16         # Untuk deteksi gambar per halaman
pytesseract==0.3.10       # OCR engine wrapper
pdf2image==1.16.3         # Konversi halaman PDF ke gambar untuk OCR
Pillow==10.3.0            # Untuk manipulasi gambar (OCR)
numpy==1.26.4             # Bantu olah data tabel OCR (jika perlu)
markdownify==0.11.6       # (optional) jika mau ubah HTML ke Markdown

# Untuk jalanin di Replit
gunicorn==21.2.0